<feat #3> KakaoMap.vue에 종속적인 카카오 지도 API를 store에 이전
